chore: Fixed alias issue by adding `baseURL`, we need to keep relative imports in the child `tsconfig.json`
