+++ package.json
@@ -48,7 +48,7 @@
     "pnpm": ">=8.0.0"
   },
   "dependencies": {
-    "@plyaz/devtools": "^1.2.39",
+    "@plyaz/devtools": "^1.3.1",
     "@tailwindcss/vite": "^4.1.8",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/user-event": "^14.6.1",
@@ -70,7 +70,6 @@
     "@eslint/markdown": "^6.5.0",
     "@nestjs/common": "^11.1.3",
     "@next/eslint-plugin-next": "^15.0.3",
-    "@plyaz/devtools": "^1.2.37",
     "@storybook/addon-a11y": "^8.6.12",
     "@storybook/addon-controls": "^8.6.12",
     "@storybook/addon-docs": "^8.6.12",
