+++ src/components/Flex/Flex.tsx
@@ -1,20 +1,20 @@
 import clsx from 'clsx';
 import React from 'react';
 
-import { Box } from '@/components/Box/Box';
-import { ALIGN_MAPPER, JUASTIFY_MAPPER } from '@/components/Grid/Grid';
+import { Box } from '@/components';
+import { ALIGN_MAPPER, JUASTIFY_MAPPER } from '@/constants/constant';
 import type { ElementType } from '@/types/type';
 
 const DIRECTION_MAPPER = {
   row: 'flex-row',
   col: 'flex-col',
-  'row-reverse': 'flex-row-reverse',
-  'col-reverse': 'flex-col-reverse',
+  row_reverse: 'flex-row-reverse',
+  col_reverse: 'flex-col-reverse',
 };
 const WRAP_MAPPER = {
   wrap: 'flex-wrap',
   nowrap: 'flex-nowrap',
-  'wrap-reverse': 'flex-wrap-reverse',
+  wrap_reverse: 'flex-wrap-reverse',
 };
 
 export interface FlexProps extends React.HTMLAttributes<HTMLElement> {
