+++ src/components/Grid/Grid.tsx
@@ -1,26 +1,9 @@
 import clsx from 'clsx';
 import React from 'react';
 
-import { Box } from '@/components/Box/Box';
+import { Box } from '@/components';
+import { ALIGN_MAPPER, JUASTIFY_MAPPER } from '@/constants/constant';
 import type { ElementType } from '@/types/type';
-
-export const JUASTIFY_MAPPER = {
-  start: 'justify-start',
-  end: 'justify-end',
-  center: 'justify-center',
-  between: 'justify-between',
-  around: 'justify-around',
-  evenly: 'justify-evenly',
-  stretch: 'justify-stretch',
-};
-
-export const ALIGN_MAPPER = {
-  start: 'items-start',
-  end: 'items-end',
-  center: 'items-center',
-  baseline: 'items-baseline',
-  stretch: 'items-stretch',
-};
 
 export interface GridProps extends React.HTMLAttributes<HTMLElement> {
   children?: React.ReactNode;
