+++ src/components/Heading/Heading.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Text } from '@/components/Text/Text';
+import { Text } from '@/components';
 
 const SIZE_MAP = {
   xs: 'xs',
