+++ src/components/Section/Section.stories.tsx
@@ -1,9 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import { expect, fn, userEvent } from '@storybook/test';
 
-import { Section } from './Section';
-
-import { Box } from '@/components/Box/Box';
+import { Box, Section } from '@/components';
 
 const meta: Meta<typeof Section> = {
   title: 'Layout/Section',
